+++ sdk/search/Azure.Search.Documents/tests/Batching/AsyncEventExtensionsTests.cs
@@ -369,7 +369,7 @@ namespace Azure.Search.Documents.Tests
         [Test]
         public async Task Cancels_StillRunning()
         {
-            TestHandler test = new TestHandler() { Delay = TimeSpan.FromMilliseconds(100) };
+            TestHandler test = new TestHandler() { Delay = TimeSpan.FromMilliseconds(500) };
             CancellationTokenSource cancellation = new CancellationTokenSource();
 
             Func<EventArgs, CancellationToken, Task> handler = null;
@@ -391,9 +391,9 @@ namespace Azure.Search.Documents.Tests
         [Test]
         public async Task Cancels_All()
         {
-            TestHandler first = new TestHandler() { Delay = TimeSpan.FromMilliseconds(100) };
-            TestHandler second = new TestHandler() { Delay = TimeSpan.FromMilliseconds(100) };
-            TestHandler third = new TestHandler() { Delay = TimeSpan.FromMilliseconds(100) };
+            TestHandler first = new TestHandler() { Delay = TimeSpan.FromMilliseconds(500) };
+            TestHandler second = new TestHandler() { Delay = TimeSpan.FromMilliseconds(500) };
+            TestHandler third = new TestHandler() { Delay = TimeSpan.FromMilliseconds(500) };
             CancellationTokenSource cancellation = new CancellationTokenSource();
 
             Func<EventArgs, CancellationToken, Task> handler = null;